decouple domain from drivers
