+++ domain/filters/filters.utils.ts
@@ -1,11 +1,11 @@
 import { WindyCiviBill } from "../types";
-import type { RepLevel } from "./filters.constants";
 import {
   AVAILABLE_TAGS,
   ChicagoTags,
+  RepLevel,
   SPONSORED_BY_REP_TAG,
   SupportedLocale,
-} from "./filters.constants";
+} from "../constants";
 
 import type {
   AddressFilter,
@@ -13,13 +13,86 @@ import type {
   Locales,
   LocationFilter,
   Nullish,
-} from "./filters.types";
+} from "../types";
+
+// TODO: Move to backend
+export const mapToReadableStatus = (
+  level: RepLevel,
+  status: string
+): { name: string; type: "in-progress" | "pass" | "fail" } => {
+  switch (level) {
+    case RepLevel.City:
+      switch (status) {
+        case "introduction":
+          return { name: "Introduced", type: "in-progress" };
+        case "referral-committee":
+          return { name: "In Committee", type: "in-progress" };
+        case "passage":
+          return { name: "Passed", type: "pass" };
+        case "substitution":
+          return { name: "Substituted", type: "in-progress" };
+        case "committee-passage-favorable":
+          return { name: "Recommended By Committee", type: "in-progress" };
+        default:
+          return {
+            name: toTitleCase(status.split("-").join(" ")),
+            type: "in-progress",
+          };
+      }
+    case RepLevel.State:
+      switch (status) {
+        case "Pass":
+          return { name: "Became Law", type: "pass" };
+        default:
+          return {
+            name: "In Progress",
+            type: "in-progress",
+          };
+      }
+    case RepLevel.National:
+      switch (status) {
+        case "Engross":
+          return { name: "Passed House", type: "in-progress" };
+        case "Enroll":
+          return {
+            name: "Awaiting Presidential Approval",
+            type: "in-progress",
+          };
+        case "Pass":
+          return { name: "Became Law", type: "pass" };
+      }
+  }
+  return { name: status, type: "in-progress" };
+};
+
+// TODO: We need to clean up the status data on the backend
+export const getLastStatus = (status: unknown): string => {
+  if (typeof status === "string") {
+    try {
+      const parsed = JSON.parse(status);
+      return parsed[parsed.length - 1];
+    } catch (e) {
+      return status;
+    }
+  }
+  if (Array.isArray(status)) {
+    return status[status.length - 1];
+  }
+  return "";
+};
+
+const toTitleCase = (str: string) => {
+  return str.replace(
+    /\w\S*/g,
+    (text) => text.charAt(0).toUpperCase() + text.substring(1).toLowerCase()
+  );
+};
 
 export const getLocale = (
-  formattedAddress: string | Nullish,
+  formattedAddress: string | Nullish
 ): null | Locales => {
   return formattedAddress && /Chicago, IL/gi.test(formattedAddress)
-    ? "Chicago"
+    ? "chicago"
     : null;
 };
 
@@ -31,7 +104,7 @@ export const getAddress = <T>(location: T): string | Nullish => {
 };
 
 export const isAddressFilter = (
-  location: unknown,
+  location: unknown
 ): location is AddressFilter => {
   if (
     typeof location === "object" &&
@@ -48,7 +121,7 @@ export const isNullish = (location: unknown | Nullish): location is Nullish => {
 };
 
 export const isSupportedLocale = (
-  locationParam: unknown,
+  locationParam: unknown
 ): locationParam is SupportedLocale => {
   if (isAddressFilter(locationParam)) {
     return false;
@@ -57,24 +130,24 @@ export const isSupportedLocale = (
     return false;
   }
   return Object.values(SupportedLocale).includes(
-    locationParam as SupportedLocale,
+    locationParam as SupportedLocale
   );
 };
 
 export const getLocation = (
-  location: string | AddressFilter | Nullish,
+  location: string | AddressFilter | Nullish
 ): string | null => {
   return isAddressFilter(location) ? location.address : location || null;
 };
 
 export const createLocationFilterFromString = (
-  locationParam: unknown,
+  locationParam: unknown
 ): LocationFilter =>
   isSupportedLocale(locationParam)
     ? locationParam
     : typeof locationParam === "string" && locationParam.length > 0
-      ? ({ address: locationParam } as AddressFilter)
-      : null;
+    ? ({ address: locationParam } as AddressFilter)
+    : null;
 
 // City Level
 export const isCityLevel = (location: LocationFilter): boolean =>
@@ -87,7 +160,7 @@ const isAddressChicago = (location: LocationFilter) =>
   isAddressFilter(location) &&
   stringIsInAddress(
     ["Chicago, IL", "Chicago,IL", "Chicago, Illinois", "Chicago,Illinois"],
-    location,
+    location
   );
 
 // State Level
@@ -103,7 +176,7 @@ export const isLocationIL = (location: LocationFilter) =>
 
 const stringIsInAddress = (variations: string[], location: AddressFilter) =>
   variations.some((str) =>
-    location.address.toLowerCase().includes(str.toLowerCase()),
+    location.address.toLowerCase().includes(str.toLowerCase())
   );
 
 export const hasTags = (tags: unknown): tags is string[] => {
@@ -111,7 +184,7 @@ export const hasTags = (tags: unknown): tags is string[] => {
 };
 
 export const getTagsBeingFiltered = (
-  filters: Pick<FilterParams, "tags" | "availableTags">,
+  filters: Pick<FilterParams, "tags" | "availableTags">
 ) => {
   return hasTags(filters?.tags) ? filters.tags : filters.availableTags;
 };
@@ -121,7 +194,7 @@ export const stringifyTags = (tags: string[]) => {
 };
 
 export const parseTagsString = (
-  tags: string | null | undefined,
+  tags: string | null | undefined
 ): string[] | null => {
   const parsed = tags?.split(",").filter((tag) => tag.length > 0);
   return hasTags(parsed) ? parsed : null;
@@ -219,4 +292,35 @@ export const findStringOverlap = (arr1: string[], arr2: string[]) => {
   }
 
   return overlap;
+};
+
+export const isLocale = (locale: unknown): locale is Locales =>
+  typeof locale === "string" &&
+  Object.values(SupportedLocale).includes(locale as SupportedLocale);
+
+export const localeValueToEnum = (locale: unknown): SupportedLocale | null => {
+  switch (locale) {
+    case "chicago":
+      return SupportedLocale.Chicago;
+    case "illinois":
+      return SupportedLocale.Illinois;
+    case "usa":
+      return SupportedLocale.USA;
+    default:
+      return null;
+  }
+};
+
+export const forEachLocale = (
+  cb: (l: SupportedLocale) => void,
+  locale?: SupportedLocale | null
+): Error | null => {
+  if (locale) {
+    cb(locale);
+  } else {
+    for (const locale of Object.values(SupportedLocale)) {
+      cb(locale);
+    }
+  }
+  return null;
 };
