--- domain/filters/index.ts
@@ -1,4 +1,0 @@
-export * from "./filters.utils";
-export * from "./filters.types";
-export * from "./filters.constants";
-export * from "./filters.selectors";
