--- domain/representatives/index.ts
@@ -1,2 +1,0 @@
-export * from "./representatives.types";
-export * from "./representatives.utils";
