+++ domain/types.ts
@@ -1,5 +1,19 @@
-import { FilteredLegislationData } from "./legislation/legislation.types";
-import { OfficialOffice } from "./representatives";
+import { RepLevel, SupportedLocale } from "./constants";
+import { OfficialOffice } from "./representatives/representatives.types";
+
+export type LegislationResult = {
+  legislation: CiviLegislationData[];
+  gpt: CiviGptLegislationData;
+};
+
+type CiviGptData = CiviGptLegislationData[keyof CiviGptLegislationData];
+
+interface FilteredLegislationData {
+  bill: CiviLegislationData;
+  gpt?: CiviGptData;
+  allTags: string[];
+  level: RepLevel;
+}
 
 // Typing the .legislation.json files
 export interface CiviLegislationData {
@@ -39,10 +53,6 @@ export interface CiviWikiLegislationData {
   tags: string[];
 }
 
-export const locales = ["chicago", "illinois", "usa"] as const;
-
-export type Locales = (typeof locales)[number];
-
 export type FeedData = {
   fullLegislation: WindyCiviBill[];
   filteredLegislation: WindyCiviBill[];
@@ -56,11 +66,30 @@ export interface WindyCiviBill extends FilteredLegislationData {
 }
 
 export interface DataStoreGetter {
-  getLegislationData: (locale: Locales) => Promise<CiviLegislationData[]>;
-  getGptLegislation: (locale: Locales) => Promise<CiviGptLegislationData>;
-  locales: typeof locales;
+  getLegislationData: (
+    locale: SupportedLocale
+  ) => Promise<CiviLegislationData[]>;
+  getGptLegislation: (
+    locale: SupportedLocale
+  ) => Promise<CiviGptLegislationData>;
+  locales: typeof SupportedLocale;
 }
 
 export type Env = {
   GOOGLE_API_KEY: string;
 };
+
+export type Locales = `${SupportedLocale}`;
+
+export type LocationFilter = SupportedLocale | AddressFilter | Nullish;
+
+export type AddressFilter = { address: string };
+
+export type Nullish = undefined | "" | null;
+
+export interface FilterParams {
+  location: LocationFilter;
+  tags: string[] | null;
+  availableTags: string[];
+  level: RepLevel | null;
+}
