+++ scraper/api/api.ts
@@ -1,4 +1,5 @@
-import { locales, Locales } from "../../domain/types";
+import { SupportedLocale } from "../../domain/constants";
+import { Locales } from "../../domain/types";
 
 export const civiLegislationApi = {
   getLegislationDataUrl: (locale: Locales): string => {
@@ -7,5 +8,5 @@ export const civiLegislationApi = {
   getGptLegislationUrl: (locale: Locales): string => {
     return `https://github.com/sartaj/windy-civi/releases/download/nightly/${locale}.legislation.gpt.json`;
   },
-  locales,
+  locales: SupportedLocale,
 };
