+++ scraper/scrapers/run.ts
@@ -1,17 +1,17 @@
-import { forEachLocale } from "../../domain/utils";
+import { forEachLocale } from "../../domain/filters/filters.utils";
 import { getLocale, getShouldSkipCache } from "../config/env";
 import { writeLegislationJSON } from "../fs/write-file";
 import { api } from "./api";
 
 const scrapeLegislation = async () => {
   const skipCache = getShouldSkipCache();
-  const locale = getLocale();
+  const localeFromEnv = getLocale();
 
   forEachLocale(async (locale) => {
     console.info("scraping for locale:", locale);
     const legislation = await api[locale]({ skipCache });
     writeLegislationJSON(locale, legislation);
-  }, locale);
+  }, localeFromEnv);
 };
 
 scrapeLegislation();
