+++ web-app/src/api/vite-api.ts
@@ -1,6 +1,5 @@
+import { SupportedLocale } from "@windycivi/domain/constants";
 import {
-  locales,
-  Locales,
   CiviLegislationData,
   CiviGptLegislationData,
   DataStoreGetter,
@@ -9,35 +8,38 @@ import {
 export const viteDataGetter: DataStoreGetter = {
   getLegislationData: (locale) => legislationApi[locale](),
   getGptLegislation: (locale) => gptApi[locale](),
-  locales,
+  locales: SupportedLocale,
 };
 
 // todo: this depends on the scraper repo to be installed, and the legislation to be there. Make this more decoupled.
-const legislationApi: Record<Locales, () => Promise<CiviLegislationData[]>> = {
-  chicago: () =>
+const legislationApi: Record<
+  SupportedLocale,
+  () => Promise<CiviLegislationData[]>
+> = {
+  [SupportedLocale.Chicago]: () =>
     import("../../../scraper/dist_legislation/chicago.legislation.json").then(
       (m) => m.default,
     ) as Promise<CiviLegislationData[]>,
-  illinois: () =>
+  [SupportedLocale.Illinois]: () =>
     import("../../../scraper/dist_legislation/illinois.legislation.json").then(
       (m) => m.default,
     ),
-  usa: () =>
+  [SupportedLocale.USA]: () =>
     import("../../../scraper/dist_legislation/usa.legislation.json").then(
       (m) => m.default,
     ) as unknown as Promise<CiviLegislationData[]>,
 };
 
-const gptApi: Record<Locales, () => Promise<CiviGptLegislationData>> = {
-  chicago: () =>
+const gptApi: Record<SupportedLocale, () => Promise<CiviGptLegislationData>> = {
+  [SupportedLocale.Chicago]: () =>
     import(
       "../../../scraper/dist_legislation/chicago.legislation.gpt.json"
     ).then((m) => m.default) as unknown as Promise<CiviGptLegislationData>,
-  illinois: () =>
+  [SupportedLocale.Illinois]: () =>
     import(
       "../../../scraper/dist_legislation/illinois.legislation.gpt.json"
     ).then((m) => m.default) as unknown as Promise<CiviGptLegislationData>,
-  usa: () =>
+  [SupportedLocale.USA]: () =>
     import("../../../scraper/dist_legislation/usa.legislation.gpt.json").then(
       (m) => m.default,
     ) as unknown as Promise<CiviGptLegislationData>,
