+++ web-app/src/app/modules/feed-ui/feed-ui.loader.tsx
@@ -2,16 +2,17 @@ import { json, type LoaderFunction } from "react-router-dom";
 import { getEnv } from "~app/modules/config";
 
 import { getFilteredLegislation } from "@windycivi/domain/api";
-import type { FilterParams } from "@windycivi/domain/filters";
+
+import { DEFAULT_FILTERS } from "@windycivi/domain/constants";
 import {
-  DEFAULT_FILTERS,
   createFilterParams,
   parseRepLevel,
-} from "@windycivi/domain/filters";
+} from "@windycivi/domain/filters/filters.utils";
+import { FilterParams } from "@windycivi/domain/types";
+import { viteDataGetter } from "../../../api/vite-api";
 import { DEFAULT_GLOBAL_STATE, RouteOption } from "./feed-ui.constants";
 import { type FeedLoaderData } from "./feed-ui.types";
 import { getCookieFromString } from "./feed-ui.utils";
-import { viteDataGetter } from "../../../api/vite-api";
 
 export const loader: LoaderFunction = async ({ request }) => {
   const globalState = DEFAULT_GLOBAL_STATE;
