+++ web-app/tsconfig.json
@@ -32,7 +32,8 @@
     "../domain/filters",
     "../domain/legislation",
     "../domain/representatives",
-    "../domain/api.ts"
+    "../domain/api.ts",
+    "../domain/constants.ts"
   ],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
